Relax credential and region requirements so they can be obtained from the host environment.
Credentials can be configured using the existing service endpoint
mechanism, or as standard environment variables
(AWS_ACCESS_KEY_ID/AWS_SECRET_ACCESS_KEY/AWS_SESSION_TOKEN) on the build host, or if
the build host is an EC2 instance, obtained from EC2 instance metadata.

Region can be configured using the existing field on each task or for an
entire pipeline using the AWS.Region variable, or as standard
environment variable on the build host (AWS_REGION) or, if the build
host is an EC2 instance, obtained from EC2 instance metadata.