Use shell for execFileSync during extension packaging (#552)
A recent node update has introduced an issue with spawnSync (execFileSync) on Windows. This fixes the issue by using a shell to run child_process execFileSync. This is because tfx is a .cmd file, which actually needs a shell. This is one of the rare cases where shell: true is necessary.